specifying one test module also works
